python/packages/kagent-rag-mcp/pyproject.toml
[project]
name = "kagent-rag-mcp"
version = "0.1.0"
description = "MCP server for querying RAG indices with dynamic tool registration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "sentence-transformers>=2.2.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "httpx>=0.25.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
kagent-rag-mcp = "kagent.rag_mcp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]
